Fix error with functionality

+++ src/services/apiCabins.js
@@ -1,6 +1,5 @@
 import supabase, { supabaseUrl } from "./supabase";
 
-// Function loads cabins from the data
 export async function getCabins() {
   const { data, error } = await supabase.from("cabins").select("*");
 
@@ -12,31 +11,41 @@ export async function getCabins() {
   return data;
 }
 
-// Function creates a new cabin
-export async function createCabin(newCabin) {
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
+
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     "/",
     ""
   );
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  // 1. Create/edit cabin
+  let query = supabase.from("cabins");
 
-  // Create cabin
-  const { data, error } = await supabase
-    .from("cabins")
-    .insert([{ ...newCabin, image: imagePath }]);
+  // A) CREATE
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+  // B) EDIT
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+
+  const { data, error } = await query.select().single();
 
   if (error) {
     console.error(error);
     throw new Error("Cabin could not be created");
   }
 
-  // Upload cabin image
+  // 2. Upload image
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
     .upload(imageName, newCabin.image);
 
-  // Delete the cabin if there was an error uploading image
+  // 3. Delete the cabin IF there was an error uplaoding image
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
     console.error(storageError);
@@ -48,9 +57,9 @@ export async function createCabin(newCabin) {
   return data;
 }
 
-// Function deletes cabin from the data
 export async function deleteCabin(id) {
   const { data, error } = await supabase.from("cabins").delete().eq("id", id);
+
   if (error) {
     console.error(error);
     throw new Error("Cabin could not be deleted");
